requirements.txt
attrs==22.2.0
Automat==22.10.0
beautifulsoup4==4.11.1
catboost==0.23
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
click==8.1.3
constantly==15.1.0
contourpy==1.0.7
cryptography==39.0.2
cssselect==1.2.0
cycler==0.11.0
faiss-cpu==1.7.0
Flask==1.1.2
fonttools==4.39.0
graphviz==0.20.1
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.0.1
Jinja2==2.11.3
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
langdetect==1.0.8
lxml==4.9.2
MarkupSafe==2.0.1
matplotlib==3.6.3
nltk==3.5
numpy==1.19.2
packaging==23.0
pandas==1.2.3
parsel==1.7.0
Pillow==9.4.0
plotly==5.13.1
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==5.4.1
queuelib==1.6.2
regex==2022.10.31
requests==2.25.1
scikit-learn==0.23.2
scipy==1.6.1
Scrapy==2.4.1
seaborn==0.12.2
service-identity==21.1.0
six==1.16.0
soupsieve==2.4
tenacity==8.2.2
threadpoolctl==3.1.0
torch==1.7.1
torchvision==0.8.2
tqdm==4.65.0
Twisted==22.10.0
typing_extensions==4.5.0
urllib3==1.26.15
w3lib==2.1.1
Werkzeug==2.0.3
zope.interface==5.5.2
